Renaming getRandomChar.js to scifitext.js to fit the theme
